implement of executor

+++ src/main/java/it/unibo/tirocinio/martelli/controller/impl/ControllerImpl.java
@@ -1,32 +1,42 @@
 package it.unibo.tirocinio.martelli.controller.impl;
 
-import java.io.IOException;
-import java.lang.reflect.InvocationTargetException;
-import java.net.URISyntaxException;
 import java.net.URL;
+import java.util.HashMap;
 import java.util.List;
-
-import it.unibo.tirocinio.martelli.controller.api.CrawlerObserver;
+import java.util.Map;
+import it.unibo.tirocinio.martelli.controller.api.ExecutorObserver;
+import it.unibo.tirocinio.martelli.controller.api.ReaderObserver;
 import it.unibo.tirocinio.martelli.controller.api.Controller;
 import it.unibo.tirocinio.martelli.database.api.Database;
 import it.unibo.tirocinio.martelli.database.impl.DatabaseSpark;
+import it.unibo.tirocinio.martelli.executor.api.Executor;
+import it.unibo.tirocinio.martelli.executor.impl.ExecutorImpl;
 import it.unibo.tirocinio.martelli.read.impl.CrawlerFactory;
+import it.unibo.tirocinio.martelli.setup.api.Setup;
+import it.unibo.tirocinio.martelli.setup.impl.SetupYml;
 
-public class ControllerImpl implements Controller, CrawlerObserver {
+public class ControllerImpl implements Controller, ReaderObserver, ExecutorObserver {
      private final Database database;
      private final CrawlerFactory factory;
-     private final URL setupPath = ClassLoader.getSystemResource("config/config.yml");
+     private final Executor executor;
 
-     public ControllerImpl() {
+     private final Map<String, Object> config;
+     private final Map<String, String> problems = new HashMap<>();
+
+     @SuppressWarnings("unchecked")
+     public ControllerImpl() throws Exception{
           database = new DatabaseSpark();
           factory = new CrawlerFactory();
+          final Setup setup = new SetupYml();
+          this.config = setup.readSetup(ClassLoader.getSystemResource("config/config.yml"));
+          executor = new ExecutorImpl(this, (List<String>) config.get("regex"));
      }
 
      @Override
-     public void execute() throws InstantiationException, IllegalAccessException, 
-                    IllegalArgumentException, InvocationTargetException, NoSuchMethodException, 
-                    SecurityException, ClassNotFoundException, IOException, URISyntaxException {
-          factory.createCrawler(setupPath, this);
+     @SuppressWarnings("unchecked")
+     public void execute() throws Exception {
+          factory.createCrawler((Map<String, Object>)config.get("crawler"), this);
+          new Thread(executor).start();
      }
 
      @Override
@@ -36,10 +46,24 @@ public class ControllerImpl implements Controller, CrawlerObserver {
 
      @Override
      public String removeDatabaseElement() {
-          return database.removeElement();
+          if (database.isNotEmpty()) {
+               return database.removeElement();
+          } else {
+               return "";
+          }
      }
 
-     public List<String> showDatabase() {
-          return database.getElements();
+     @Override
+     public void addProblem(final String url, final String value) {
+          problems.put(url, value);
+     }
+
+     public String showProblems() {
+          return problems.toString();
+     }
+
+     @Override
+     public String showDatabse() {
+          return database.getElements().toString();
      }
 }
